faasta_types

Struct Header

#[repr(C)]
pub struct Header {
    pub name: String,
    pub value: String,
}

Fields
name: Stringvalue: String
Implementations

impl Header
where Struct<FieldPair<String, String>>: IStable, String: IStable,

pub const fn has_optimal_layout() -> bool

Returns true if the layout for Header is smaller or equal to that Rust would have generated for it.